quad heur
leaf split de quadratic split

+++ src/dadat1/Rectangle.java
@@ -1,12 +1,12 @@
 package dadat1;
 
 public class Rectangle {
-  int x1;
-  int x2;
-  int y1;
-  int y2;
+  double x1;
+  double x2;
+  double y1;
+  double y2;
   
-  public Rectangle(int x1, int x2, int y1, int y2) {
+  public Rectangle(double x1, double x2, double y1, double y2) {
     this.x1 = x1;
     this.x2 = x2;
     this.y1 = y1;
@@ -15,7 +15,7 @@ public class Rectangle {
   
   public Rectangle surroundRect(Rectangle rect) {
     //entrega el mbr de this y rect
-    int rx1, rx2, ry1, ry2;
+    double rx1, rx2, ry1, ry2;
     rx1 = Math.min(this.x1, rect.x1);
     rx2 = Math.max(this.x2, rect.x2);
     ry1 = Math.min(this.y1, rect.y1);
@@ -24,8 +24,8 @@ public class Rectangle {
     return res;
   }
   
-  public int giveArea() {
-    int res = (x2-x1)*(y2-y1);
+  public double giveArea() {
+    double res = (x2-x1)*(y2-y1);
     return res;
   }
 }
